我的作业页面编写，具体逻辑待完善

+++ LeaderDiandu/LeaderDiandu/Model/HBTestWorkManager.h
@@ -11,7 +11,7 @@
 @interface HBTestWorkManager : NSObject
 
 @property (nonatomic, strong)NSArray *workArray;
-@property (nonatomic, assign)NSInteger selIndex;
+@property (nonatomic, assign)long selIndex;
 
 - (void)parseTestWork:(NSString *)path;
